Add live view prototype
This is a small prototype of using live view. It demonstrates live view
using a small counter and a width. The counter information will be
broadcast back to all clients.

The intent of this is to grow to support the messaging aspects of the
application. As messages are stored, they will be shared back to their
corresponding subscriptions.